refactor(server): type admin context, trim order payload, split errors
S6 — every admin sub-app now declares the context the admin middleware
actually guarantees, so 47 `as JWTPayload` casts are gone. One shared
`AdminEnv` in src/types/hono.ts, applied to all 12 sub-apps and to the
middleware that sets the variable, so producer and consumers are pinned
to one declaration instead of drifting. Casts survive only in 11 test
fixtures that build deliberately partial payloads.

S8 — GET /admin/orders/:id stops shipping what nothing reads. The
handlerLogs relation and its three user joins are dropped (the table is
still written; only the read had no consumer), complaints is capped to
the one row its unique index already allows, and the catalog service row
is narrowed to id+name — which also stops sending the shop's cost base
and list price to every browser that opens an order. statusLogs stays
uncapped: two screens render the full timeline and print its count.

S13 — the two modules both named `errors` now say what they hold:
src/errors.ts -> src/http-exceptions.ts, src/utils/errors.ts ->
src/utils/pg-error.ts, with the errorHandler suite moved out to its own
error-handler.test.ts. 52 import sites updated; git recorded all three
as renames, so history follows.
