+++ packages/server/src/modules/campaigns/campaign-eligibility.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "bun:test";
-import { BadRequestException } from "@/errors";
+import { BadRequestException } from "@/http-exceptions";
 import { assertCampaignUsable } from "@/modules/campaigns/campaign.service";
 import { campaignIneligibilityReason } from "@/schema/campaign-eligibility";
