+++ packages/server/src/modules/orders/order-payment.service.test.ts
@@ -1,5 +1,5 @@
 import { beforeEach, describe, expect, it, mock } from "bun:test";
-import { BadRequestException, ForbiddenException } from "@/errors";
+import { BadRequestException, ForbiddenException } from "@/http-exceptions";
 import { captureRejection } from "@/test-support/capture-rejection";
 import type { JWTPayload } from "@/types";
