+++ packages/server/src/modules/orders/order.repository.ts
@@ -7,7 +7,7 @@ import {
   ordersServicesTable,
   ordersTable,
 } from "@/db/schema";
-import { BadRequestException } from "@/errors";
+import { BadRequestException } from "@/http-exceptions";
 import type { NormalizedOrderListQuery } from "@/modules/orders/order.schema";
 import {
   deriveOrderRefundStatus,
