+++ packages/server/src/modules/permissions/permissions.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "bun:test";
-import { ForbiddenException } from "@/errors";
+import { ForbiddenException } from "@/http-exceptions";
 import {
   assertCanCancelOrderService,
   assertCanCreateOrder,
