+++ packages/server/src/routes/admin/payment-methods.ts
@@ -1,6 +1,6 @@
 import { Hono } from "hono";
 import { StatusCodes } from "http-status-codes";
-import { NotFoundException } from "@/errors";
+import { NotFoundException } from "@/http-exceptions";
 import {
   GETPaymentMethodsQuerySchema,
   POSTPaymentMethodSchema,
@@ -13,10 +13,11 @@ import {
   updatePaymentMethod,
 } from "@/modules/payment-methods/payment-method.service";
 import { idParamSchema } from "@/schema/param";
+import type { AdminEnv } from "@/types/hono";
 import { success } from "@/utils/http";
 import { zodValidator } from "@/utils/zod-validator-wrapper";
 
-const app = new Hono()
+const app = new Hono<AdminEnv>()
   .get("/", zodValidator("query", GETPaymentMethodsQuerySchema), async (c) => {
     const query = c.req.valid("query");
     const paymentMethods = await getPaymentMethods(query);
