+++ packages/server/src/routes/admin/products.ts
@@ -1,6 +1,6 @@
 import { Hono } from "hono";
 import { StatusCodes } from "http-status-codes";
-import { NotFoundException } from "@/errors";
+import { NotFoundException } from "@/http-exceptions";
 import {
   POSTProductSchema,
   PUTProductSchema,
@@ -12,10 +12,11 @@ import {
   updateProduct,
 } from "@/modules/products/product.service";
 import { idParamSchema } from "@/schema/param";
+import type { AdminEnv } from "@/types/hono";
 import { success } from "@/utils/http";
 import { zodValidator } from "@/utils/zod-validator-wrapper";
 
-const app = new Hono()
+const app = new Hono<AdminEnv>()
   .get("/", async (c) => {
     const products = await getProducts();
