+++ packages/server/src/routes/admin/shifts.ts
@@ -10,13 +10,13 @@ import {
   getCurrentShift,
   getShifts,
 } from "@/modules/shifts/shift.service";
-import type { JWTPayload } from "@/types";
+import type { AdminEnv } from "@/types/hono";
 import { success } from "@/utils/http";
 import { zodValidator } from "@/utils/zod-validator-wrapper";
 
-const app = new Hono()
+const app = new Hono<AdminEnv>()
   .get("/", zodValidator("query", GETShiftsQuerySchema), async (c) => {
-    const user = c.get("jwtPayload") as JWTPayload;
+    const user = c.get("jwtPayload");
     const query = c.req.valid("query");
 
     const { items, meta } = await getShifts(user, query);
@@ -24,12 +24,12 @@ const app = new Hono()
     return c.json(success(items, undefined, meta));
   })
   .get("/current", async (c) => {
-    const user = c.get("jwtPayload") as JWTPayload;
+    const user = c.get("jwtPayload");
     const shift = await getCurrentShift(user);
     return c.json(success(shift ?? null));
   })
   .post("/clock-in", zodValidator("json", POSTClockInSchema), async (c) => {
-    const user = c.get("jwtPayload") as JWTPayload;
+    const user = c.get("jwtPayload");
     const body = c.req.valid("json");
 
     const shift = await clockIn({ user, storeId: body.store_id });
@@ -40,7 +40,7 @@ const app = new Hono()
     );
   })
   .post("/clock-out", async (c) => {
-    const user = c.get("jwtPayload") as JWTPayload;
+    const user = c.get("jwtPayload");
     const shift = await clockOut(user);
     return c.json(success(shift, "Clocked out successfully"));
   });
