+++ packages/server/src/routes/admin/stores.ts
@@ -1,6 +1,6 @@
 import { Hono } from "hono";
 import { StatusCodes } from "http-status-codes";
-import { NotFoundException } from "@/errors";
+import { NotFoundException } from "@/http-exceptions";
 import {
   GETNearestStoreQuerySchema,
   PATCHStoreSchema,
@@ -16,10 +16,11 @@ import {
   updateStoreStatus,
 } from "@/modules/stores/store.service";
 import { idParamSchema } from "@/schema/param";
+import type { AdminEnv } from "@/types/hono";
 import { success } from "@/utils/http";
 import { zodValidator } from "@/utils/zod-validator-wrapper";
 
-const app = new Hono()
+const app = new Hono<AdminEnv>()
   .get("/", async (c) => {
     const stores = await getStores();
