+++ packages/server/src/routes/auth.ts
@@ -5,7 +5,7 @@ import { sign } from "hono/jwt";
 import { StatusCodes } from "http-status-codes";
 import { db } from "@/db";
 import { usersTable } from "@/db/schema";
-import { ForbiddenException, UnauthorizedException } from "@/errors";
+import { ForbiddenException, UnauthorizedException } from "@/http-exceptions";
 import type { JWTPayload } from "@/types/jwt";
 import { success } from "@/utils/http";
 import { zodValidator } from "@/utils/zod-validator-wrapper";
