+++ packages/server/src/types/hono.ts
@@ -1,0 +1,6 @@
+import type { JwtVariables } from "hono/jwt";
+import type { JWTPayload } from "./jwt";
+
+export interface AdminEnv {
+  Variables: JwtVariables<JWTPayload>;
+}
