+++ packages/server/src/utils/pg-error.test.ts
@@ -1,0 +1,85 @@
+import { describe, expect, it } from "bun:test";
+import { StatusCodes } from "http-status-codes";
+import { mapPostgresError, redactDetail } from "@/utils/pg-error";
+
+const PHONE = "+628123456789";
+
+describe("mapPostgresError", () => {
+  it("tells the counter the customer is already on file without printing the other customer's number", () => {
+    // Two branches take the same walk-in on the same day. Postgres reports the
+    // clash by quoting the number it found, and that number belongs to whoever
+    // registered first — it must not reach the screen of whoever tried second.
+    const { message, status } = mapPostgresError({
+      code: "23505",
+      constraint: "customers_phone_number_unique",
+      detail: `Key (phone_number)=(${PHONE}) already exists.`,
+    });
+
+    expect(message).toBe("A customer with that phone number already exists");
+    expect(message).not.toContain(PHONE);
+    expect(status).toBe(StatusCodes.CONFLICT);
+  });
+
+  it("says the same thing whether the shop is on the dev or the production database", () => {
+    // Dev names the constraint customers_phone_number_key, production names it
+    // customers_phone_number_unique — both databases are live.
+    const { message } = mapPostgresError({
+      code: "23505",
+      constraint: "customers_phone_number_key",
+    });
+
+    expect(message).toBe("A customer with that phone number already exists");
+  });
+
+  it("names the mistake when a refund is written for more than the customer paid", () => {
+    const { message, status } = mapPostgresError({
+      code: "23514",
+      constraint: "refunded_amount_valid_check",
+      detail: `Failing row contains (91, ORD-0007, 1, 2, 50000, 80000, Budi Santoso, ${PHONE}).`,
+    });
+
+    expect(message).toBe("Refund cannot be more than the amount paid");
+    expect(message).not.toContain("Budi Santoso");
+    expect(status).toBe(StatusCodes.BAD_REQUEST);
+  });
+
+  it("keeps the half-typed order off the screen when a required field was left blank", () => {
+    const { message, status } = mapPostgresError({
+      code: "23502",
+      detail: `Failing row contains (91, ORD-0007, null, Budi Santoso, ${PHONE}).`,
+    });
+
+    expect(message).toBe("A required field is missing");
+    expect(message).not.toContain("Budi Santoso");
+    expect(status).toBe(StatusCodes.BAD_REQUEST);
+  });
+
+  it("stays quiet about a database failure nobody has mapped", () => {
+    const { message, status } = mapPostgresError({
+      code: "40001",
+      detail: `Key (phone_number)=(${PHONE}) already exists.`,
+    });
+
+    expect(message).toBe("Something went wrong");
+    expect(message).not.toContain(PHONE);
+    expect(status).toBe(StatusCodes.INTERNAL_SERVER_ERROR);
+  });
+});
+
+describe("redactDetail", () => {
+  it("keeps which column clashed and drops whose number it was", () => {
+    expect(redactDetail(`Key (phone_number)=(${PHONE}) already exists.`)).toBe(
+      "Key (phone_number)=(…) already exists."
+    );
+  });
+
+  it("drops the whole customer row Postgres echoes back on a rejected save", () => {
+    const redacted = redactDetail(
+      `Failing row contains (91, Budi Santoso, ${PHONE}, budi@example.com).`
+    );
+
+    expect(redacted).toBe("Failing row contains (…).");
+    expect(redacted).not.toContain(PHONE);
+    expect(redacted).not.toContain("Budi Santoso");
+  });
+});
