+++ packages/server/src/utils/zod-validator-wrapper.ts
@@ -2,7 +2,7 @@ import { zValidator as zv } from "@hono/zod-validator";
 import type { ValidationTargets } from "hono/types";
 import { z } from "zod";
 import type { $ZodType } from "zod/v4/core";
-import { BadRequestException } from "@/errors";
+import { BadRequestException } from "@/http-exceptions";
 
 export const zodValidator = <
   T extends $ZodType,
